Move data coverage table to its own component
