+++ src/components/DataCoverageTable/DataCoverageTable.stories.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import CoverageTable, {
-  COVERAGE_SUMMARY,
-} from 'screens/DataApi/DataCoverageTable';
+import CoverageTable, { COVERAGE_SUMMARY } from 'components/DataCoverageTable';
 
 export default {
   title: 'Data API/Data Availability Table',
