+++ src/components/DataCoverageTable/index.ts
@@ -1,0 +1,8 @@
+import CoverageTable, {
+  COVERAGE_SUMMARY,
+  FIELDS_PRETTY_NAMES,
+} from './DataCoverageTable';
+
+export { COVERAGE_SUMMARY, FIELDS_PRETTY_NAMES };
+
+export default CoverageTable;
